+++ components/common/form/checkFormFieldErrors.ts
@@ -1,6 +1,6 @@
 import type { FormFieldType } from '@charmverse/core/prisma-client';
 
-import type { FormFieldInput } from '../../../lib/forms/interfaces';
+import type { FormFieldInput } from 'lib/forms/interfaces';
 
 import { nonDuplicateFieldTypes } from './constants';
