+++ components/common/form/fields/CharmEditorInputField.tsx
@@ -2,7 +2,7 @@ import { useTheme } from '@emotion/react';
 
 import { CharmEditor } from 'components/common/CharmEditor';
 import { FieldWrapper } from 'components/common/form/fields/FieldWrapper';
-import type { ControlFieldProps, FieldProps } from 'lib/forms/interfaces';
+import type { ControlFieldProps, FieldProps } from 'components/common/form/interfaces';
 import type { PageContent } from 'lib/prosemirror/interfaces';
 
 type Props = ControlFieldProps & FieldProps & { multiline?: boolean; rows?: number };
