+++ components/common/form/fields/ImageField.tsx
@@ -1,8 +1,7 @@
 import { Button } from 'components/common/Button';
+import type { ControlFieldProps, FieldProps } from 'components/common/form/interfaces';
 import ImageSelector from 'components/common/ImageSelector/ImageSelector';
 import type { PageContent } from 'lib/prosemirror/interfaces';
-
-import type { ControlFieldProps, FieldProps } from '../../../../lib/forms/interfaces';
 
 import type { FieldWrapperProps } from './FieldWrapper';
 import { FieldWrapper } from './FieldWrapper';
