+++ components/common/form/fields/PersonInputField.tsx
@@ -1,7 +1,7 @@
 import { forwardRef } from 'react';
 
 import { FieldWrapper } from 'components/common/form/fields/FieldWrapper';
-import type { ControlFieldProps, FieldProps } from 'lib/forms/interfaces';
+import type { ControlFieldProps, FieldProps } from 'components/common/form/interfaces';
 
 import { InputSearchMemberMultiple } from '../InputSearchMember';
